Started working on assignment

+++ lib/bloc_record/persistence.rb
@@ -95,19 +95,31 @@ module Persistence
       true
     end
 
-    def destroy_all(conditions_hash=nil)
-      if conditions_hash && !conditions_hash.empty?
-        conditions_hash = BlocRecord::Utility.convert_keys(conditions_hash)
-        conditions = conditions_hash.map {|key, value| "#{key}=#{BlocRecord::Utility.sql_strings(value)}"}.join(" and ")
+    def destroy_all(*conditions)
+      conditions
+      case conditions.first
+      when String
+        if conditions.length > 1
+          # Answer to question #3
+        else
+          # Answer to question #2
+        end
 
-        connection.execute <<-SQL
-          DELETE FROM #{table}
-          WHERE #{conditions}
-        SQL
-      else
-        connection.execute <<-SQL
-          DELETE FROM #{table}
-        SQL
+      when Hash
+        conditions_hash = conditions.first
+        if conditions_hash && !conditions_hash.empty?
+          conditions_hash = BlocRecord::Utility.convert_keys(conditions_hash)
+          conditions = conditions_hash.map {|key, value| "#{key}=#{BlocRecord::Utility.sql_strings(value)}"}.join(" and ")
+
+          connection.execute <<-SQL
+            DELETE FROM #{table}
+            WHERE #{conditions}
+          SQL
+        else
+          connection.execute <<-SQL
+            DELETE FROM #{table}
+          SQL
+        end
       end
 
       true
